Move WeightGradStore into execution.py and FP8 weight caching into operators
