+++ pithtrain/dualpipe/__init__.py
@@ -1,9 +1,6 @@
 from pithtrain.dualpipe.dualpipev import DualPipeV, Microbatch
-from pithtrain.dualpipe.utils import FP8WeightCacheControl, WeightGradStore
 
 __all__ = [
     "DualPipeV",
-    "FP8WeightCacheControl",
     "Microbatch",
-    "WeightGradStore",
 ]
